Change getTasksInBoundingBox to return review info usernames (#647)
* Change ClusteredPoint PointReview to use longs for reviewedBy
  and reviewRequestedBy (these should have been longs before)

* When fetching getTasksInBoundingBox insert extra username info
  into the ClusterPoint.pointReview field for reviewedBy and
  reviewRequestedBy

+++ app/org/maproulette/controllers/api/ChallengeController.scala
@@ -187,11 +187,11 @@ class ChallengeController @Inject()(override val childController: TaskController
       Json.toJson(List[JsValue]())
     } else {
       val mappers = Some(this.dalManager.user.retrieveListById(-1, 0)(tasks.map(
-        t => t.pointReview.reviewRequestedBy.getOrElse(0).toLong)).map(u =>
+        t => t.pointReview.reviewRequestedBy.getOrElse(0L))).map(u =>
           u.id -> Json.obj("username" -> u.name, "id" -> u.id)).toMap)
 
       val reviewers = Some(this.dalManager.user.retrieveListById(-1, 0)(tasks.map(
-        t => t.pointReview.reviewedBy.getOrElse(0).toLong)).map(u =>
+        t => t.pointReview.reviewedBy.getOrElse(0L))).map(u =>
           u.id -> Json.obj("username" -> u.name, "id" -> u.id)).toMap)
 
       val jsonList = tasks.map { task =>
